+++ src/domain/mod.rs
@@ -1,1 +1,7 @@
 pub mod tenant;
+pub mod user;
+
+#[allow(unused_imports)]
+pub use tenant::*;
+#[allow(unused_imports)]
+pub use user::*;